# openapi.yaml
openapi: 3.1.0
info:
  title: Open ClubManager REST API
  version: 2.0.0
  summary: Versioned REST API for Open ClubManager
  description: |
    OpenAPI specification for the currently implemented REST API under `/api/v2`.

    Notes:
    - The legacy `/api/*` v1 endpoints were removed.
    - Authentication uses bearer access tokens for protected endpoints.
    - Refresh/logout use an HTTP-only cookie named `refreshToken`.
servers:
  - url: http://localhost:3001/api/v2
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          examples: [INTERNAL_ERROR, MEMBER_VERSION_CONFLICT]
        error:
          type: string
      required: [error]
      additionalProperties: false
...
    DatabaseMode:
      type: string
      enum: [sqlite-local, mysql-shared]
    SetupStatus:
      type: object
      properties:
        setupRequired:
          type: boolean
        userCount:
          type: integer
        databaseConfigured:
          type: boolean
        databaseMode:
          $ref: '#/components/schemas/DatabaseMode'
      required: [setupRequired, userCount, databaseConfigured, databaseMode]
      additionalProperties: false
    ConfigureDatabaseRequest:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/DatabaseMode'
        databaseUrl:
          type: string
      required: [mode]
      additionalProperties: false
    ConfigureDatabaseResponse:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/DatabaseMode'
        databaseUrl:
          type: string
      required: [mode, databaseUrl]
      additionalProperties: false
    DatabaseSettingsRequest:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/DatabaseMode'
        databaseUrl:
          type: string
      required: [mode, databaseUrl]
      additionalProperties: false
    SetupInitializeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [email, password]
      additionalProperties: false
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
      additionalProperties: false
...
    UserPreferenceRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, key, value, updatedAt]
      additionalProperties: false
    UserPreferenceValueResponse:
      type: object
      properties:
        key:
          type: string
        value:
          description: Parsed JSON value stored in the preference entry.
        updatedAt:
          type: string
          format: date-time
      required: [key, value, updatedAt]
      additionalProperties: false
    UpsertUserPreferenceRequest:
      type: object
      properties:
        key:
          type: string
        value:
          description: Any JSON-serializable value.
      required: [key, value]
      additionalProperties: false
    UpdateUserPreferenceRequest:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required: [key, value]
      additionalProperties: false
...
    UpdateAppPreferenceRequest:
      type: object
      properties:
        value:
          description: Any JSON-serializable value.
      required: [value]
      additionalProperties: false
    ValidationRequest:
      type: object
      properties:
        path:
          type: string
      required: [path]
      additionalProperties: false
    ValidationDatabaseUrlRequest:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/DatabaseMode'
        databaseUrl:
          type: string
      required: [mode, databaseUrl]
      additionalProperties: false
    ValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        i18nToken:
          type: string
      required: [valid]
      additionalProperties: false
    SetDbPathRequest:
      type: object
      properties:
        dbPath:
          type: string
      required: [dbPath]
      additionalProperties: false
    SetDbPathResponse:
      type: object
      properties:
        status:
          type: string
        dbPath:
          type: string
      required: [status]
      additionalProperties: false
    DbPathResponse:
      type: object
      properties:
        dbPath:
          type:
            - string
            - 'null'
      required: [dbPath]
      additionalProperties: false
    DatabaseSettings:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/DatabaseMode'
        databaseUrl:
          type: string
      required: [mode, databaseUrl]
      additionalProperties: false
...
  responses:
...
    BadRequest:
      description: Invalid request payload or query parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Authenticated but not allowed, or refresh token invalid/expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected server-side error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /setup/status:
    get:
      tags: [setup]
      summary: Get setup status
      operationId: getSetupStatus
      responses:
        '200':
          description: Current setup status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupStatus'
        '500':
          $ref: '#/components/responses/InternalError'
  /setup/configure-database:
    post:
      tags: [setup]
      summary: Configure the initial database mode and URL
      operationId: configureSetupDatabase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigureDatabaseRequest'
      responses:
        '200':
          description: Database configuration stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigureDatabaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /setup/initialize:
    post:
      tags: [setup]
      summary: Create the initial admin user
      operationId: initializeSetup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupInitializeRequest'
      responses:
        '201':
          description: Initial admin user created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /preferences:
    get:
      tags: [preferences]
      summary: List stored user preferences
      operationId: listPreferences
      responses:
        '200':
          description: User preferences.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPreferenceRecord'
    post:
      tags: [preferences]
      summary: Create or update a user preference for the implicit user
      operationId: upsertPreference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertUserPreferenceRequest'
      responses:
        '201':
          description: Preference stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferenceRecord'
...
  /preferences/app/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [preferences]
      summary: Get one app preference by key
      operationId: getAppPreferenceByKey
      responses:
        '200':
          description: Single app preference value.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [preferences]
      summary: Update one app preference by key
      operationId: updateAppPreferenceByKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppPreferenceRequest'
      responses:
        '200':
          description: Updated app preference response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
  /preferences/{identifier}:
    parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
        description: |
          For GET this is the preference key.
          For PUT this is the numeric preference id.
    get:
      tags: [preferences]
      summary: Get a user preference by key
      operationId: getPreferenceByKey
      responses:
        '200':
          description: Preference value with parsed JSON.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferenceValueResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [preferences]
      summary: Update a user preference by id
      operationId: updatePreferenceById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPreferenceRequest'
      responses:
        '200':
          description: Updated preference record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferenceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
  /validation/check-db-path:
    post:
      tags: [validation]
      summary: Validate a SQLite database path
      operationId: validateDbPath
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
      responses:
        '200':
          description: Validation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Missing path.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
  /validation/check-db-url:
    post:
      tags: [validation]
      summary: Validate a database URL for the selected mode
      operationId: validateDbUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationDatabaseUrlRequest'
      responses:
        '200':
          description: Validation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Missing or invalid mode.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
  /settings/set-db-path:
    post:
      tags: [settings]
      summary: Set active database path
      operationId: setDbPath
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetDbPathRequest'
      responses:
        '200':
          description: Database path updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetDbPathResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/db-path:
    get:
      tags: [settings]
      summary: Get active database path
      operationId: getDbPath
      responses:
        '200':
          description: Current active database path.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DbPathResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/database:
    get:
      tags: [settings]
      summary: Get database mode and connection URL
      operationId: getDatabaseSettings
      responses:
        '200':
          description: Current database settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseSettings'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [settings]
      summary: Update database mode and connection URL
      operationId: updateDatabaseSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseSettingsRequest'
      responses:
        '200':
          description: Database settings updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigureDatabaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...